+++ src/container/__tests__/module.test.ts
@@ -1,77 +1,110 @@
-import { Observable, of } from "rxjs";
 import { Container, ContainerError } from "../container";
-import { IModuleDependencies, Module, ModuleError } from "../module";
+import { IModuleDependencies, IModuleDestroy, IModuleHook, Module, ModuleError } from "../module";
 
 // Tests for Module up/down order.
 const moduleUpOrder: number[] = [];
+const moduleUpHookOrder: number[] = [];
 const moduleDownOrder: number[] = [];
-let moduleDestroy = 0;
+const moduleDownHookOrder: number[] = [];
+const moduleDestroyOrder: number[] = [];
+const moduleDestroyHookOrder: number[] = [];
 
 class Test1 extends Module {
   public static readonly moduleName: string = "Test1";
-  public moduleUp(): void {
-    moduleUpOrder.push(1);
+  public moduleUp(...args: IModuleHook[]) {
+    return super.moduleUp(...args, async () => {
+      moduleUpOrder.push(1);
+    });
   }
-  public moduleDown(): void {
-    moduleDownOrder.push(1);
+  public moduleDown(...args: IModuleHook[]) {
+    return super.moduleDown(...args, async () => {
+      moduleDownOrder.push(1);
+    });
   }
-  public moduleDestroy(): void {
-    moduleDestroy++;
+  public moduleDestroy(...args: IModuleDestroy[]) {
+    return super.moduleDestroy(...args, () => {
+      moduleDestroyOrder.push(1);
+    });
   }
 }
 
 class Test2 extends Module {
   public static readonly moduleName: string = "Test2";
-  public moduleDependencies(...prev: IModuleDependencies[]): IModuleDependencies {
+  public moduleDependencies(...prev: IModuleDependencies[]) {
     return super.moduleDependencies(...prev, { test1: Test1, test3: Test3 });
   }
-  public moduleUp(): Observable<void> {
-    moduleUpOrder.push(2);
-    return of(undefined);
+  public moduleUp(...args: IModuleHook[]) {
+    return super.moduleUp(...args, async () => {
+      moduleUpOrder.push(2);
+    });
   }
-  public moduleDown(): Observable<void> {
-    moduleDownOrder.push(2);
-    return of(undefined);
+  public moduleDown(...args: IModuleHook[]) {
+    return super.moduleDown(...args, async () => {
+      moduleDownOrder.push(2);
+    });
   }
-  public moduleDestroy(): void {
-    moduleDestroy++;
+  public moduleDestroy(...args: IModuleDestroy[]) {
+    return super.moduleDestroy(...args, () => {
+      moduleDestroyOrder.push(2);
+    });
   }
 }
 
 class Test3 extends Module {
   public static readonly moduleName: string = "Test3";
   public readonly test1!: Test1;
-  public moduleDependencies(...prev: IModuleDependencies[]): IModuleDependencies {
+  public moduleDependencies(...prev: IModuleDependencies[]) {
     return super.moduleDependencies(...prev, { test1: Test1 });
   }
-  public async moduleUp(): Promise<void> {
-    moduleUpOrder.push(3);
+  public moduleUp(...args: IModuleHook[]) {
+    return super.moduleUp(...args, async () => {
+      moduleUpOrder.push(3);
+      moduleUpHookOrder.push(3);
+    });
   }
-  public async moduleDown(): Promise<void> {
-    moduleDownOrder.push(3);
+  public moduleDown(...args: IModuleHook[]) {
+    return super.moduleDown(...args, async () => {
+      moduleDownOrder.push(3);
+      moduleDownHookOrder.push(3);
+    });
   }
-  public moduleDestroy(): void {
-    moduleDestroy++;
+  public moduleDestroy(...args: IModuleDestroy[]) {
+    return super.moduleDestroy(...args, () => {
+      moduleDestroyOrder.push(3);
+      moduleDestroyHookOrder.push(3);
+    });
   }
 }
 
 class Test4 extends Test3 {
   public static readonly moduleName: string = "Test4";
   public readonly test2!: Test2;
-  public moduleDependencies(...prev: IModuleDependencies[]): IModuleDependencies {
+  public moduleDependencies(...prev: IModuleDependencies[]) {
     return super.moduleDependencies(...prev, { test2: Test2 });
   }
-  public async moduleUp(): Promise<void> {
-    moduleUpOrder.push(4);
+  public moduleUp(...args: IModuleHook[]) {
+    return super.moduleUp(...args, async () => {
+      moduleUpOrder.push(4);
+      moduleUpHookOrder.push(4);
+    });
   }
-  public async moduleDown(): Promise<void> {
-    moduleDownOrder.push(4);
+  public moduleDown(...args: IModuleHook[]) {
+    return super.moduleDown(...args, async () => {
+      moduleDownOrder.push(4);
+      moduleDownHookOrder.push(4);
+    });
+  }
+  public moduleDestroy(...args: IModuleDestroy[]) {
+    return super.moduleDestroy(...args, () => {
+      moduleDestroyOrder.push(4);
+      moduleDestroyHookOrder.push(4);
+    });
   }
 }
 
 class DependencyErrorModule extends Module {
   public static readonly moduleName: string = "DependencyErrorModule";
-  public moduleDependencies(...prev: IModuleDependencies[]): IModuleDependencies {
+  public moduleDependencies(...prev: IModuleDependencies[]) {
     return super.moduleDependencies(...prev, { test2: Test2 });
   }
 }
@@ -80,18 +113,21 @@ describe("Module", () => {
   const CONTAINER = new Container("Test1").registerModules([Test1, Test2, Test3, Test4]);
 
   it("up calls moduleUp methods in expected order", async () => {
-    await CONTAINER.up().toPromise();
-    expect(moduleUpOrder).toEqual([1, 3, 2, 4]);
+    await CONTAINER.up();
+    expect(moduleUpOrder).toEqual([1, 3, 2, 3, 4]);
+    expect(moduleUpHookOrder).toEqual([3, 3, 4]);
   });
 
   it("down calls moduleDown methods in expected order", async () => {
-    await CONTAINER.down().toPromise();
-    expect(moduleDownOrder).toEqual([4, 2, 3, 1]);
+    await CONTAINER.down();
+    expect(moduleDownOrder).toEqual([4, 3, 2, 3, 1]);
+    expect(moduleDownHookOrder).toEqual([4, 3, 3]);
   });
 
   it("destroy calls moduleDestroy methods", () => {
     CONTAINER.destroy();
-    expect(moduleDestroy).toEqual(4);
+    expect(moduleDestroyOrder).toEqual([1, 2, 3, 4, 3]);
+    expect(moduleDestroyHookOrder).toEqual([3, 4, 3]);
   });
 
   it("registerModules throws error for duplicate module names", (done) => {
@@ -110,25 +146,25 @@ describe("Module", () => {
     expect(t4.test2 instanceof Test2).toEqual(true);
   });
 
-  it("unknown module dependency throws error during container.up", (done) => {
-    const container = new Container("test").registerModule(DependencyErrorModule);
-    container.up().subscribe({
-      next: () => done.fail(),
-      error: (error) => {
-        expect(error instanceof ModuleError).toEqual(true);
-        done();
-      }
-    });
+  it("unknown module dependency throws error during container.up", async (done) => {
+    try {
+      const container = new Container("test").registerModule(DependencyErrorModule);
+      await container.up();
+      done.fail();
+    } catch (error) {
+      expect(error instanceof ModuleError).toEqual(true);
+      done();
+    }
   });
 
-  it("unknown module dependency throws error during container.down", (done) => {
-    const container = new Container("test").registerModule(DependencyErrorModule);
-    container.down().subscribe({
-      next: () => done.fail(),
-      error: (error) => {
-        expect(error instanceof ModuleError).toEqual(true);
-        done();
-      }
-    });
+  it("unknown module dependency throws error during container.down", async (done) => {
+    try {
+      const container = new Container("test").registerModule(DependencyErrorModule);
+      await container.down();
+      done.fail();
+    } catch (error) {
+      expect(error instanceof ModuleError).toEqual(true);
+      done();
+    }
   });
 });