+++ src/lib/modules/__tests__/logs.test.ts
@@ -20,11 +20,11 @@ describe("Logs", () => {
   const LOGS = CONTAINER.resolve<TestLogs>(TestLogs.moduleName);
 
   beforeAll(async () => {
-    await CONTAINER.up().toPromise();
+    await CONTAINER.up();
   });
 
   afterAll(async () => {
-    await CONTAINER.down().toPromise();
+    await CONTAINER.down();
     CONTAINER.destroy();
   });