+++ src/lib/modules/__tests__/metrics.test.ts
@@ -20,11 +20,11 @@ describe("Metric", () => {
   const METRICS = CONTAINER.resolve<TestMetrics>(TestMetrics.name);
 
   beforeAll(async () => {
-    await CONTAINER.up().toPromise();
+    await CONTAINER.up();
   });
 
   afterAll(async () => {
-    await CONTAINER.down().toPromise();
+    await CONTAINER.down();
     CONTAINER.destroy();
   });